+++ contrib/JMeshLib/src/PRIMITIVES/heap.cpp
@@ -56,8 +56,8 @@ int abstractHeap::upheap(int k)
   heap[fk] = t;
   if (positions != NULL)
   {
-   positions[(j_voidint)f] = k;
-   positions[(j_voidint)t] = fk;
+   positions[(j_voidint)(size_t)f] = k;
+   positions[(j_voidint)(size_t)t] = fk;
   }
   return upheap(fk);
  }
@@ -81,8 +81,8 @@ int abstractHeap::downheap(int k)
   heap[j] = t;
   if (positions != NULL)
   {
-   positions[(j_voidint)f] = k;
-   positions[(j_voidint)t] = j;
+   positions[(j_voidint)(size_t)f] = k;
+   positions[(j_voidint)(size_t)t] = j;
   }
   return downheap(j);
  }
@@ -95,18 +95,18 @@ int abstractHeap::insert(void *t)
  if (numels == maxels) return -1;
 
  heap[++numels] = t;
- if (positions != NULL) positions[(j_voidint)t] = numels;
+ if (positions != NULL) positions[(j_voidint)(size_t)t] = numels;
  return upheap(numels);
 }
 
 void *abstractHeap::removeHead()
 {
  void *t = heap[1];
- if (positions != NULL) positions[(j_voidint)t] = 0;
+ if (positions != NULL) positions[(j_voidint)(size_t)t] = 0;
  heap[1] = heap[numels--];
  if (numels)
  {
-  if (positions != NULL) positions[(j_voidint)heap[1]] = 1;
+  if (positions != NULL) positions[(j_voidint)(size_t)heap[1]] = 1;
   downheap(1);
  }
